+++ pyproject.toml
@@ -57,8 +57,7 @@ packages = [
 
 [tool.poetry.plugins."neptune.extensions"]
 "experimental_remote_signals" = "neptune_experimental.remote_signals:initialize"
-"experimental_operation_error_processor" = "neptune_experimental.operation_error_processor:initialize"
-"experimental_incremental_batch_size" = "neptune_experimental.incremental_batch_size:initialize"
+"experimental_async_operation_processor" = "neptune_experimental.custom_async_operation_processor:initialize"
 
 [tool.poetry.urls]
 "Tracker" = "https://github.com/neptune-ai/neptune-client-experimental/issues"
